workaround xcode linking bug
